add filter in SupportOBSTWL
