+++ src/components/ListWords/index.js
@@ -1,0 +1,2 @@
+export { default as ListLinks } from './ListLinks';
+export { default as ListWords } from './ListWords';
