+++ src/components/ListWords/style.js
@@ -1,8 +1,10 @@
 import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles((theme) => ({
-  twl: {
-    color: theme.palette.text.disabled,
+  linkContainer: {
+    cursor: 'pointer',
+    padding: theme.spacing(1 / 2, 2),
+    display: 'block',
   },
 }));
 export default useStyles;
