+++ src/components/SupportTWL/index.js
@@ -1,3 +1,1 @@
 export { default as SupportTWL } from './SupportTWL';
-export { default as ListLink } from './ListLink';
-export { default as ListWords } from './ListWords';
