+++ src/hooks/index.js
@@ -1,2 +1,6 @@
-export * from './useWindowSize';
+export * from './useChangeColorTWL';
+export * from './useChunk';
+export * from './useListWordsReference';
 export * from './useScrollToVerse';
+export * from './useSelectTypeUniqueWords';
+export * from './useWindowSize';
